Code restructure with FreeShippingCards.
This is so that when users remove all the shipping rates, the OfferFreeShippingCard and MinimumOrderCard will not be rendered.
