+++ js/src/components/flat-shipping-rates-input-cards/flat-shipping-rates-input-cards.js
@@ -1,52 +1,12 @@
 /**
- * External dependencies
- */
-import { useState } from '@wordpress/element';
-
-/**
  * Internal dependencies
  */
 import EstimatedShippingRatesCard from './estimated-shipping-rates-card';
-import MinimumOrderCard from './minimum-order-card';
-import OfferFreeShippingCard from './offer-free-shipping-card';
+import FreeShippingCards from './free-shipping-cards';
 
 const FlatShippingRatesInputCards = ( props ) => {
 	const { audienceCountries, value, onChange = () => {} } = props;
-
-	const displayOfferFreeShippingCard = value.some( ( el ) => el.rate > 0 );
-
-	let initialOfferFreeShippingValue;
-	if ( displayOfferFreeShippingCard ) {
-		initialOfferFreeShippingValue = value.some(
-			( el ) => el.options?.free_shipping_threshold > 0
-		);
-	}
-
-	const [ offerFreeShippingValue, setOfferFreeShippingValue ] = useState(
-		initialOfferFreeShippingValue
-	);
-
-	const handleOfferFreeShippingChange = ( newOfferFreeShippingValue ) => {
-		/**
-		 * When users select 'no', we remove all the
-		 * shippingRate.options.free_shipping_threshold.
-		 */
-		if ( ! newOfferFreeShippingValue ) {
-			const newValue = value.map( ( el ) => {
-				return {
-					...el,
-					options: {
-						...el.options,
-						free_shipping_threshold: undefined,
-					},
-				};
-			} );
-
-			onChange( newValue );
-		}
-
-		setOfferFreeShippingValue( newOfferFreeShippingValue );
-	};
+	const displayFreeShippingCards = value.some( ( el ) => el.rate > 0 );
 
 	return (
 		<>
@@ -55,14 +15,8 @@ const FlatShippingRatesInputCards = ( props ) => {
 				value={ value }
 				onChange={ onChange }
 			/>
-			{ displayOfferFreeShippingCard && (
-				<OfferFreeShippingCard
-					value={ offerFreeShippingValue }
-					onChange={ handleOfferFreeShippingChange }
-				/>
-			) }
-			{ offerFreeShippingValue && (
-				<MinimumOrderCard value={ value } onChange={ onChange } />
+			{ displayFreeShippingCards && (
+				<FreeShippingCards value={ value } onChange={ onChange } />
 			) }
 		</>
 	);
